requirements.txt
pycryptodome
PyQt5
playwright
pystray 
pillow